tests/wpt/web-platform-tests/css/selectors/missing-right-token.html

<!DOCTYPE html>
<html>
<head>
  <meta id="expected" charset="utf-8">
  <title>Selectors: attribute selectors with missing right token</title>
  <link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
  <link rel="help" href="https://drafts.csswg.org/selectors-4/#attribute-selectors">
  <link rel="help" href="https://drafts.csswg.org/css-syntax/#rule-defs">
  <meta name="assert" content="Checks attribute selectors with missing right token are matched.">
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>
<body>


  <div id="container">
    <span></span>
    <span class="cls"></span>
  </div>

  <script>
    test(() => {
      assert_equals(document.querySelector('meta[charset="utf-8"'), expected);
      assert_equals(document.querySelector('meta[charset="utf-8'), expected);

      assert_equals(container.querySelectorAll('span:not([class]').length, 1);
      assert_equals(container.querySelectorAll('span:not([class').length, 1);
    }, "attribute selectors with missing right tokens succeed");
  </script>
</body>
</html>
